Reorganization of ClientApp
