+++ ClientApp/src/components/DisplayHelpers/Interface/LabelledItems/ColumnarLabelledText.tsx
