+++ ClientApp/src/components/DisplayHelpers/Interface/LabelledItems/LabelledText.scss
@@ -1,4 +1,4 @@
-@import "../../variables";
+@import "../../../../variables";
 
 .LabelledText {
   .Label {
