--- ClientApp/src/components/DisplayHelpers/Interface/Pagination/Pagination.scss
@@ -1,3 +1,0 @@
-.Pagination {
-  
-}
