+++ ClientApp/src/components/Employees/EmployeeDetail.tsx
@@ -7,11 +7,11 @@ import { Link, RouteComponentProps } from 'react-router-dom'
 
 import { requestJSONWithErrorHandler } from '../../helpers/requestHelpers'
 import ContentWrapper from '../Wrappers/ContentWrapper'
-import CLText from '../DisplayHelpers/ColumnarLabelledText'
+import CLText from '../DisplayHelpers/Interface/LabelledItems/ColumnarLabelledText'
 import Date from '../DisplayHelpers/FormattedDate'
-import Address from '../DisplayHelpers/Address'
+import Address from './Address'
 import EditableAddress from './EditableAddress'
-import LabelledText from '../DisplayHelpers/LabelledText'
+import LabelledText from '../DisplayHelpers/Interface/LabelledItems/LabelledText'
 import TimelineEntryList from './TimelineEntryList'
 import AddComment from './AddComment'
 import EditableStringField from './EditableStringField'
