+++ ClientApp/src/components/TaskLogEntries/TaskOutcome.tsx
