+++ ClientApp/src/components/Wrappers/ErrorPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import LabelledText from '../DisplayHelpers/LabelledText'
+import LabelledText from '../DisplayHelpers/Interface/LabelledItems/LabelledText'
 
 const smallCodeWrapper = (content: string) => (
   <code>
