+++ ClientApp/src/helpers/objectHelper.ts
@@ -1,10 +1,3 @@
-import { NullableString } from '../types/NullableString'
-import moment from 'moment'
-
-export function createDateAsUTCFromString(dateString: string): Date {
-  return moment.utc(dateString).toDate()
-}
-
 export function undefinedIfNull<T>(obj: T | null): T | undefined {
   return obj === null ? undefined : obj
 }
@@ -15,8 +8,4 @@ export function nullIfUndefined<T>(obj: T | undefined): T | null {
 
 export function emptyStringIfNull<T>(obj: T | null): T | string {
   return obj === null ? '' : obj
-}
-
-export function dateOrUndefined(dateString: NullableString): Date | undefined {
-  return dateString ? createDateAsUTCFromString(dateString) : undefined
 }
