+++ ClientApp/src/index.tsx
@@ -1,10 +1,10 @@
 import 'reflect-metadata'
-import './App.scss'
+import './components/App.scss'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
 import { Provider } from 'react-redux'
-import App from './App'
+import App from './components/App'
 import { unregister } from './registerServiceWorker'
 import { OidcProvider } from 'redux-oidc'
 import userManager from './store/utils/userManager'
